waig-rag-poc/requirements.txt
# Core RAG Dependencies
openai>=1.12.0
langfuse>=2.0.0
qdrant-client>=1.7.0
llama-parse>=0.3.3
sentence-transformers>=2.2.2

pydantic>=2,<3
pydantic-settings>=2.3,<3
python-dotenv>=1.0.0

# Document Processing
pypdf2>=3.0.1
pdfplumber>=0.10.0
python-docx>=1.0.1
markdown>=3.5.1

# Data Processing & ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tiktoken>=0.5.0

# API & Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
aiofiles>=23.2.0
httpx>=0.26.0
tenacity>=8.0.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional but recommended
redis>=5.0.0  # For caching
celery>=5.3.0  # For background tasks 
